% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{access_on_groupe}
\alias{access_on_groupe}
\title{calcul de l'accessibilité par compartiments}
\usage{
access_on_groupe(
  groupe,
  ou_4326,
  quoi_4326,
  routing,
  k,
  tmax,
  opp_var,
  ttm_out,
  pids,
  dir,
  t2d
)
}
\arguments{
\item{groupe}{groupe}

\item{ou_4326}{les positions.}

\item{quoi_4326}{les aménités/opportunités.}

\item{routing}{le système de routing.}

\item{k}{facteur de subsampling}

\item{tmax}{temps maximal des isochrones.}

\item{opp_var}{liste des variables d'aménités/opportunités.}

\item{ttm_out}{switch selon lequel on performe l'aggrégation ou on garde la matrice des temps}

\item{pids}{liste des process en //}

\item{dir}{dossier des résultats intermédiaires.}

\item{t2d}{enregistre ou non sur disque les résultats}
}
\description{
calcul de l'accessibilité par compartiments
}
