% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_utils.r
\name{choisir_jour_transit}
\alias{choisir_jour_transit}
\title{Choisit le jour du transit}
\usage{
choisir_jour_transit(plage, start_time = "8:00:00")
}
\arguments{
\item{plage}{une plage}

\item{start_time}{l'heure}
}
\value{
un jour au format date
}
\description{
Permet de choisir un jour de semaine (ni samedi ni dimanche dans une plage
au milieu de la plage
}
