% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_utils.r
\name{choix_jour}
\alias{choix_jour}
\title{Choisit un jour}
\usage{
choix_jour(directory)
}
\arguments{
\item{directory}{le répertoire où se trouve le GTFS}
}
\value{
une date
}
\description{
En prenant un GTFS en entrée, la fonction choisit parmi les jours de semaine
un jour qui existe pour tous les réseaux
}
