% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_cutoff.R
\name{cumulative_time_cutoff}
\alias{cumulative_time_cutoff}
\title{Cumulative access based on a travel cost cutoff}
\usage{
cumulative_time_cutoff(
  data,
  cutoff,
  opportunity_col,
  travel_cost_col = "travel_time",
  by_col
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel cost matrix in long format,
containing the columns with origin ids, destination ids, the travel cost
between each origin-destination pair (this can be either in  terms of
travel time, distance or monetary costs), and number of population and
opportunities located in each origin and destination.}

\item{cutoff}{A \code{numeric} value indicating the maximum travel time considered.}

\item{opportunity_col}{A string. The name of the column in \code{land_use_data}
with the number of opportunities/resources/services to be considered when
calculating accessibility levels.}

\item{travel_cost_col}{A string. The name of the column in \code{travel_matrix}
with the travel cost between origins and destinations. Defaults to
\code{"travel_time"}.}

\item{by_col}{Either \code{NULL} (the default) or a \code{character} vector. When not
\code{NULL}, indicates the \code{travel_matrix} columns that should be used to group
the accessibility estimates by. For example, if \code{travel_matrix} includes a
departure time column, that specifies the departure time of each entry in
the data frame, passing \code{"departure_time"} to this argument results in
accessibility estimates grouped by origin and by departure time.}
}
\value{
A \code{data.table} containing the accessibility estimate for each origin
in the travel matrix.
}
\description{
Calculates the number of opportunities accessible under a given specified
travel cost cutoff.

This function is generic over any kind of numeric travel cost,
such as distance, time and money.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# Active accessibility: number of schools accessible from each origin
df <- cumulative_time_cutoff(data = ttm,
                            cutoff = 30,
                            opportunity_col = 'schools',
                            travel_cost_col = 'travel_time',
                            by_col = 'from_id')
head(df)

# Passive accessibility: number of people that can reach each destination
df <- cumulative_time_cutoff(data = ttm,
                            cutoff = 30,
                            opportunity_col = 'population',
                            travel_cost_col = 'travel_time',
                            by_col = 'to_id')
head(df)
}
\seealso{
Other Cumulative access: 
\code{\link{cumulative_time_interval}()}
}
\concept{Cumulative access}
