% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_interval.R
\name{cumulative_time_interval}
\alias{cumulative_time_interval}
\title{Cumulative access based on time interval}
\usage{
cumulative_time_interval(
  data = df,
  start = 20,
  end = 30,
  by_col = "from_id",
  summary = "mean"
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format}

\item{start}{An \code{integer} indicating the \code{start} point of the travel time interval}

\item{end}{An \code{integer} indicating the \code{end} point of the travel time interval}

\item{by_col}{A string pointing to the name of the column of origin or destination}

\item{summary}{A \code{string} indicating the summary measure to be used. It accepts
either \code{mean} (Default) or \code{median}.}
}
\value{
A \code{data.table} object.
}
\description{
The function calculates the average or median number of opportunities
accessible within a travel time interval specified by the user.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_poa.csv", package = "accessibility")
ttm <- read.csv(data_path)

df <- cumulative_time_interval(data = ttm,
                              start = 20,
                              end = 30,
                              by_col = 'from_id',
                              summary ='mean')
head(df)

}
\seealso{
Other Cumulative access: 
\code{\link{cumulative_time_threshold}()}
}
\concept{Cumulative access}
