% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_cutoff.R
\name{cumulative_time_threshold}
\alias{cumulative_time_threshold}
\title{Cumulative access based on time threshold}
\usage{
cumulative_time_threshold(data, cutoff = 20, by_col = "from_id")
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format}

\item{cutoff}{An \code{numeric}}

\item{by_col}{A string pointing to the name of the column of origin or destination}
}
\value{
A \code{data.table} object.
}
\description{
The function calculates the number of opportunities accessible under a given
travel time threshold specified by the user.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_poa.csv", package = "accessibility")
ttm <- read.csv(data_path)

# Active accessibility: number of schools accessible from each origin
df <- cumulative_time_threshold(data = ttm,
                               cutoff = 30,
                               by_col = 'from_id')
head(df)

# Passive accessibility: number of people that can reach each destination
df <- cumulative_time_threshold(data = ttm,
                               cutoff = 30,
                               by_col = 'to_id')
}
\seealso{
Other Cumulative access: 
\code{\link{cumulative_time_interval}()}
}
\concept{Cumulative access}
