% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_exponential.R
\name{decay_exponential}
\alias{decay_exponential}
\title{Negative exponential decay function}
\usage{
decay_exponential(decay_value)
}
\arguments{
\item{decay_value}{A \code{numeric}. The calibration parameter that, when
multiplied by the travel cost, is used as the exponent of \code{e} in the
negative exponential function.}
}
\value{
A \code{function} that takes a generic travel cost (\code{numeric}) as an input
and returns an impedance factor (\code{numeric}).
}
\description{
Returns a negative exponential impedance function  to be used inside
accessibility calculating functions.

This function is generic over any kind of numeric travel cost,
such as distance, time and money.
}
\examples{
impedance <- decay_exponential(decay_value = 0.1)

impedance(20)

impedance(35)

}
\seealso{
Other Impedance functions: 
\code{\link{decay_binary}()},
\code{\link{decay_linear}()},
\code{\link{decay_power}()}
}
\concept{Impedance functions}
