% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_linear.R
\name{decay_linear}
\alias{decay_linear}
\title{Linear decay function}
\usage{
decay_linear(cutoff)
}
\arguments{
\item{cutoff}{A \code{numeric}. A number indicating the travel cost cutoff until
which the impedance factor decays linearly. From this point onward the
impedance factor is equal to 0.}
}
\value{
A \code{function} that takes a generic travel cost (\code{numeric}) as an input
and returns an impedance factor (\code{numeric}).
}
\description{
Returns a linear impedance function to be used inside accessibility
calculating functions.

This function is generic over any kind of numeric travel cost,
such as distance, time and money.
}
\examples{
impedance <- decay_linear(cutoff = 30)

impedance(20)

impedance(35)

}
\seealso{
Other decay functions: 
\code{\link{decay_binary}()},
\code{\link{decay_exponential}()},
\code{\link{decay_power}()}
}
\concept{decay functions}
