% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_logistic.R
\name{decay_logistic}
\alias{decay_logistic}
\title{Logistic-CDF decay function}
\usage{
decay_logistic(median, sd)
}
\arguments{
\item{median}{A \code{numeric} vector. The median or inflection point
of the logistic CDF in minutes of travel time.}

\item{sd}{A \code{numeric} vector with same length as \code{median}.
The standard deviation in minutes of the logistic-CDF decay function
must be greater than 0 and less than 120.
Smaller values result in more rapid decay.}
}
\value{
A \code{function} that takes a generic travel cost vector (\code{numeric}) as
an input and returns a vector of weights (\code{numeric}).
}
\description{
Returns a logist-cdf weighting function parameterized with the
median (inflection point) and standard deviation to be used inside
accessibility calculating functions.

This function is generic over any kind of numeric travel cost,
such as distance, time and money.
}
\details{
When using a function created with \code{decay_logistic()}, the
output is named after the combination of median (\code{"m"}) and sd (\code{"s"})
\itemize{
\item e.g. given the median \code{c(10, 20)} and the sd \code{c(10, 20)},
the outputs will be named \code{"m(10);s(10)"}, \code{"m(20);s(20)"}.
}
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
weighting_function <- decay_logistic(
  median = seq(10, 120, by = 10),
  sd = 10
)

weighting_function(seq(0, 120, by = 5))

weighting_function <- decay_logistic(
  c(10, 10, 10, 10, 20, 20, 20, 20),
  c(2, 4, 6, 8, 10, 12, 2, 4, 6, 8, 10, 12)
)

weighting_function(seq(0, 120, by = 5))
\dontshow{\}) # examplesIf}
}
\seealso{
Other decay functions: 
\code{\link{decay_binary}()},
\code{\link{decay_exponential}()},
\code{\link{decay_linear}()},
\code{\link{decay_power}()},
\code{\link{decay_stepped}()}
}
\concept{decay functions}
