% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{delayRouting}
\alias{delayRouting}
\title{définit le type du routeur}
\usage{
delayRouting(delay, routing)
}
\arguments{
\item{delay}{délai pour l'heure de départ}

\item{routing}{système de routing}
}
\description{
définit le type du routeur
}
