% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr router.R
\name{dgr_distances_by_paires}
\alias{dgr_distances_by_paires}
\title{Full distance}
\usage{
dgr_distances_by_paires(
  idINSes,
  com2com,
  routeur,
  path = "dgr",
  chunk = 5000000L,
  overwrite = TRUE
)
}
\arguments{
\item{idINSes}{une table d'idINS, avec les colonnes idINS, from (lgl), to (lgl) COMMUNE, DCLT}

\item{com2com}{une table donnant les paires de relation}

\item{routeur}{un routeur dodgr, généré par routing_setup_dodgr}

\item{path}{chemin d'enregistrement du dataset}

\item{chunk}{taille du paquet de découpage}

\item{overwrite}{écrase des précédents

pour chaque paire COMMUNE, DCLT le routage est fait sur le produit des
from et to indiqués pour cette paire}
}
\value{
rien, mais un dataset est enregistré dans path
}
\description{
Alternative à iso acessibilité, plus simple
ne renvoie que les distances, temps et dzplus
exécute le calcul par paires de COMMUNE/DCLT
certains batchs sont assez petits
}
