% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr router.R
\name{dgr_distances_full}
\alias{dgr_distances_full}
\title{Full distance}
\usage{
dgr_distances_full(
  idINSes,
  routeur,
  path = "dgr",
  overwrite = TRUE,
  chunk = 5000000L
)
}
\arguments{
\item{idINSes}{une table d'idINS, avec les colonnes idINS, from (lgl), to (lgl) COMMUNE, DCLT}

\item{routeur}{un routeur dodgr, généré par routing_setup_dodgr

pour chaque paire from et to}
}
\value{
un dataset
}
\description{
Alternative à iso acessibilité, plus simple
ne renvoie que les distances, temps et dzplus
exécute le calcul sur toutes les paires
de façon à avoir des gros batchs
}
