% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr router.R
\name{dgr_onedistance}
\alias{dgr_onedistance}
\title{dgr_onedistance}
\usage{
dgr_onedistance(routeur, from, to, parallel = TRUE)
}
\arguments{
\item{routeur}{le routeur préprocessé}

\item{from}{les points de départ (préindexés sur le routeur$graph)}

\item{to}{les points d'arrivée (préindexés sur le routeur$graph)}
}
\value{
un data.table
}
\description{
Calcule un data frame contenant time_travel, distance et dzplus
}
