% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr router.R
\name{dodgr_pairs}
\alias{dodgr_pairs}
\title{ttm sur des paires}
\usage{
dodgr_pairs(od, routing, chunk = Inf)
}
\arguments{
\item{od}{tibble origine destination}

\item{routing}{routing (voir dodgr_routing_setup)}

\item{chunk}{nombre de paires demandées d'un coup}
}
\value{
a tibble
}
\description{
ttm sur des paires
}
