% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr router.R
\name{download_osmsc}
\alias{download_osmsc}
\title{Télécharge d'OSM pour dodgr en format silicate}
\usage{
download_osmsc(box, workers = 1, .progress = TRUE)
}
\arguments{
\item{box}{les limites de la zone à télécharger, au format st_bbox()}

\item{workers}{le nombre de workers}

\item{.progress}{affiche un indicateur de progression}
}
\value{
un osmdata_sc
}
\description{
le format silicate permet à dodgr de pondérer les tournants,
les arrêts aux feux rouge ainsi que les restrictions de circulation
}
