% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{dt_ttm}
\alias{dt_ttm}
\title{calcul du temps de trajet avec data.table (??)}
\usage{
dt_ttm(o, d, tmax, routing)
}
\arguments{
\item{o}{origine}

\item{d}{destination}

\item{tmax}{temps max pour le trajet}

\item{routing}{système de routing}
}
\description{
calcul du temps de trajet avec data.table (??)
}
