% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floating_catchment_area.R
\name{floating_catchment_area}
\alias{floating_catchment_area}
\title{Floating catchment area accessibilitt}
\usage{
floating_catchment_area(
  data,
  fca_metric,
  orig_col,
  dest_col,
  population_col,
  opportunity_col,
  decay_function,
  cutoff = NULL,
  decay_value = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{fca_metric}{A \code{string} indicating which floating cacthment area
accessibility measure to use. Options avaiblable: \code{"2SFCA"} and \code{"BFCA"}.}

\item{orig_col}{A \code{string} with the name of the column of origin ids.}

\item{dest_col}{A \code{string} with the name of the column of destination ids.}

\item{population_col}{A \code{string} with the name of the column of origin with
population count.}

\item{opportunity_col}{A \code{string} with the name of the column of destination
with  the number of opportunities / resources / services.}

\item{decay_function}{A string. Which decay function to use when calculating
accessibility. One of step, exponential, fixed_exponential, linear
or logistic. Please see the details to understand how each
alternative works and how they relate to the \code{cutoffs} and
\code{decay_value} parameters.}

\item{cutoff}{A numeric vector. This parameter has different effects for each
decay function: it indicates the cutoff times in minutes when
calculating cumulative opportunities accessibility with the
\code{step} function...}

\item{decay_value}{A number. Extra parameter to be passed to the selected
\code{decay_function}. Has no effects when \code{decay_function} is either
\code{step} or \code{exponential}.}
}
\value{
A \code{data.table} object.
}
\description{
Calculates accessibility accouting for the competition of resources using one
of the multiple accessibility metrics in the floating catchment area family.
The function currently includes 2SFCA and BFCA
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# 2SFCA with a step decay function
df <- floating_catchment_area(
       data = ttm,
       fca_metric = '2SFCA',
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = 'step',
       cutoff = 30
       )
head(df)

# BFCA with a step decay function
df <- floating_catchment_area(
       data = ttm,
       fca_metric = 'BFCA',
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = 'step',
       cutoff = 30
       )
head(df)

df2 <- floating_catchment_area(data = ttm,
        fca_metric = '2SFCA',
        orig_col = 'from_id',
        dest_col = 'to_id',
        opportunity_col = 'jobs',
        population_col = 'population',
        decay_function = 'negative_exponential',
        decay_value = 0.5
        )

head(df2)

}
\seealso{
Other Floating catchment area: 
\code{\link{fca_2sfca}()},
\code{\link{fca_bfca}()}
}
\concept{Floating catchment area}
