% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_plot.R
\name{get_line}
\alias{get_line}
\title{Extrait une ligne d'un GTFS}
\usage{
get_line(gtfs, route_id)
}
\arguments{
\item{gtfs}{le gtfs que l'on veut analyser}

\item{route_id}{l'id de la route}
}
\value{
un sf représentant la ligne en 3035
}
\description{
Extrait une ligne d'un GTFS
}
