% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{get_routing}
\alias{get_routing}
\title{récupère les infos sur le routeur}
\usage{
get_routing(routing, groupe)
}
\arguments{
\item{routing}{système de routing.}

\item{groupe}{??}
}
\description{
récupère les infos sur le routeur
}
