% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r5 router.R
\name{get_setup_r5}
\alias{get_setup_r5}
\title{récupère le setup du système de routing r5}
\usage{
get_setup_r5(
  data_path,
  verbose = FALSE,
  temp_dir = FALSE,
  use_elevation = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{data_path}{path}

\item{verbose}{par défaut, FALSE}

\item{temp_dir}{par défaut, FALSE}

\item{use_elevation}{par défaut, FALSE}

\item{overwrite}{par défaut, FALSE}
}
\description{
récupère le setup du système de routing r5
}
