% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini_index}
\alias{gini_index}
\title{Gini index}
\usage{
gini_index(accessibility_data, sociodemographic_data, opportunity, population)
}
\arguments{
\item{accessibility_data}{A data frame. The accessibility levels whose
inequality should be calculated. Must contain the columns \code{id} and any
others specified in \code{opportunity}.}

\item{sociodemographic_data}{A data frame. The distribution of
sociodemographic characteristics of the population in the study area cells.
Must contain the columns \code{id} and any others specified in \code{population} and
\code{income}.}

\item{opportunity}{A string. The name of the column in \code{accessibility_data}
with the accessibility levels to be considered when calculating inequality
levels.}

\item{population}{A string. The name of the column in \code{sociodemographic_data}
with the number of people in each cell. Used to weigh accessibility levels
when calculating inequality.}
}
\value{
A numeric value with the Gini index estimate for the study area.
}
\description{
Calculates the Gini index of a given accessibility distribution.
}
\examples{
data_dir <- system.file("extdata", package = "accessibility")
travel_matrix <- readRDS(file.path(data_dir, "travel_matrix.rds"))
land_use_data <- readRDS(file.path(data_dir, "land_use_data.rds"))

access <- cumulative_cutoff(
  travel_matrix = travel_matrix,
  land_use_data = land_use_data,
  cutoff = 30,
  opportunity = "jobs",
  travel_cost = "travel_time"
)

gini <- gini_index(
  accessibility_data = access,
  sociodemographic_data = land_use_data,
  opportunity = "jobs",
  population = "population"
)
gini

}
\concept{Inequality}
