% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_model.R
\name{gravity_access}
\alias{gravity_access}
\title{Gravity-based accessibility measures}
\usage{
gravity_access(
  data,
  opportunity_col,
  by_col,
  decay_function,
  cutoff = NULL,
  decay_value = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{opportunity_col}{A \code{string} indicating the name of the column with
data on the opportunities to be considered.}

\item{by_col}{A \code{string} with the name of the column of origin or
destination that should be considered, indicating whether accessibility
levels should by calculated by each origin (active accessibility) or
destination (passive accessibility).}

\item{decay_function}{A string. Which decay function to use when calculating
accessibility. One of step, exponential, fixed_exponential, linear
or logistic. Please see the details to understand how each
alternative works and how they relate to the \code{cutoffs} and
\code{decay_value} parameters.}

\item{cutoff}{A numeric vector. This parameter has different effects for each
decay function: it indicates the cutoff times in minutes when
calculating cumulative opportunities accessibility with the
\code{step} function...}

\item{decay_value}{A number. Extra parameter to be passed to the selected
\code{decay_function}. Has no effects when \code{decay_function} is either
\code{step} or \code{exponential}.}
}
\value{
A \code{data.table} object.
}
\description{
The function calculates gravity-based accessibility measures...
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df_linear <- gravity_access(data = ttm,
                           opportunity_col = 'schools',
                           by_col = 'from_id',
                           decay_function = 'linear',
                           cutoff = 30)

head(df_linear)

df_neg_exp <- gravity_access(data = ttm,
                            opportunity_col = 'schools',
                            by_col = 'from_id',
                            decay_function = 'negative_exponential',
                            decay_value = 0.5)
head(df_neg_exp)

}
\seealso{
Other Cumulative access: 
\code{\link{cumulative_time_cutoff}()},
\code{\link{cumulative_time_interval}()}
}
\concept{Cumulative access}
