% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.r
\name{idINS2square}
\alias{idINS2square}
\title{Crée un pavage carré en tant que géométrie d'un sf à partir de idINS.}
\usage{
idINS2square(ids, resolution = NULL)
}
\arguments{
\item{ids}{vecteur d'idINS.}

\item{resolution}{resolution, par défaut celle attachée à idINS.}
}
\description{
Crée un pavage carré en tant que géométrie d'un sf à partir de idINS.
}
