% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impedance_fun.R
\name{impedance_fun}
\alias{impedance_fun}
\title{General impedance functions}
\usage{
impedance_fun(t_ij, decay_function, cutoff = NULL, decay_value = NULL)
}
\arguments{
\item{t_ij}{A number indicating the travel cost (time or money) between
origin (i) and destination (j).}

\item{decay_function}{A string. Which decay function to use when calculating
accessibility. One of step, exponential, fixed_exponential, linear
or logistic. Please see the details to understand how each
alternative works and how they relate to the \code{cutoffs} and
\code{decay_value} parameters.}

\item{cutoff}{A numeric vector. This parameter has different effects for each
decay function: it indicates the cutoff times in minutes when
calculating cumulative opportunities accessibility with the
\code{step} function...}

\item{decay_value}{A number. Extra parameter to be passed to the selected
\code{decay_function}. Has no effects when \code{decay_function} is either
\code{step} or \code{exponential}.}
}
\value{
A \code{numeric} impedance factor
}
\description{
Provides a numer of impedance functions to be used inside \code{accessibility} functions.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_poa.csv", package = "accessibility")
ttm <- read.csv(data_path)

# Active accessibility: number of schools accessible from each origin
df <- cumulative_time_cutoff(data = ttm,
                            opportunity_colname = 'schools',
                            cutoff = 30,
                            by_colname = 'from_id')
head(df)

# Passive accessibility: number of people that can reach each destination
df <- cumulative_time_cutoff(data = ttm,
                            opportunity_colname = 'population',
                            cutoff = 30,
                            by_colname = 'to_id')
head(df)
}
\concept{Impedance functions}
