% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso2time.r
\name{iso2time}
\alias{iso2time}
\title{Transforme des isochrones en temps d'accès}
\usage{
iso2time(isoraster, seuils = median(raster::cellStats(isoraster, median)))
}
\arguments{
\item{isoraster}{le raster en entrée}

\item{seuils}{les seuils pour la sortie}
}
\value{
un raster avec autant de couches que d'éléments dans seuils
}
\description{
Cette fonction transforme un raster avec des couches donnant pour des temps donnés des opportunités atteintes
}
