% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_accessibilite.r
\name{iso_accessibilite}
\alias{iso_accessibilite}
\title{Calcul de l'accessibilité}
\usage{
iso_accessibilite(
  quoi,
  ou = NULL,
  res_quoi = Inf,
  resolution = ifelse(is.null(ou), 200, Inf),
  var_quoi = "individuals",
  routing = "r5",
  tmax = 10L,
  pdt = 1L,
  chunk = 2e+07,
  future = FALSE,
  out = ifelse(is.finite(resolution), resolution, "raster"),
  ttm_out = FALSE,
  logs = ".",
  dir = NULL,
  table2disk = if (is.null(dir)) FALSE else TRUE
)
}
\arguments{
\item{quoi}{sf de variables numériques de différentes aménités, qui vont être agrégées au sein d'isochrones.
Différentes options sont possibles pour le format et les formes passées en entrée.}

\item{ou}{positions sur lesquelles sont calculées les accessibilités.
Si NULL, un raster est défini par défaut en ayant la même empreinte que les oportunités, à la résolution resolution.}

\item{res_quoi}{rasterisation éventuelle des lieux à la résolution donnée. Par défaut, reste un sf.}

\item{resolution}{résolution retenue pour l'accessibilité. Par défaut, 200m si ou est NULL, Inf sinon.}

\item{var_quoi}{si rasterisation, définie la fonction d'agrégation. Non implémenté.}

\item{routing}{défini le moteur de routage. Le routing dopit être initialisé par un setup_routing_*}

\item{tmax}{temps maximal pour les calculs d'isochrone. Par défaut, 10 mn.}

\item{pdt}{pas de temps pour le calcul des isochrones. Par defaut, incrément de 1mn.}

\item{chunk}{taille des paquets pour le compartimentage des calculs.}

\item{future}{calcul parallélisé. Par défaut, FALSE.}

\item{out}{format de sortie ("data.table", "sf", "raster"). Par défaut, raster.}

\item{ttm_out}{conserver travel_time_matrix. Par défaut, FALSE. Peut générer un objet volumineux.}

\item{logs}{dossier pour le fichier log, par defaut la variable localdata (non définie).}

\item{dir}{dossier des étapes du calcul (permet de récupérer ce qui a déjà calculé en cas de plantage).}

\item{table2disk}{enregistre les étapes du calcul. Par défaut, si dir est défini, TRUE.}
}
\value{
Par défaut, la sortie est une liste de bricks (ensemble de rasters). Chaque élément de la liste renvoie à une aménité et la brick donne tous les rasters selon le temps de déplacement.
}
\description{
\code{iso_accessibilite} calcule l'accessibilité à des aménités définies sur une carte (sous forme de sf) et
renvoie un raster à une certaine résolution. Le calcul des distances isochroniques est organisé par secteurs et
est majoré pour évacuer des calculs inutiles entre secteurs éloignés.
}
\details{
Les étapes du calcul peuvent être enregistrées pour ne pas repartir de zéro si le calcul plante en cours de route.
l'algorithme fonctionne comme cela :
\enumerate{
\item découpe les groupes d'origines
\item par groupe
\item prend un point au hasard
\item calcule les distances entre ce point et les cibles
\item calcule les distance entre le point et les autres du paquet
\item sélectionne pour chaque auter point du paquet les cibles atteignables
\item calcule les distances
\item aggrege
\item cumule
\item rasterize
}
}
