% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{iso_ouetquoi_4326}
\alias{iso_ouetquoi_4326}
\title{construit la grille pour projetter le résultat en crs 4326
\code{iso_ouetquoi} projette sur 4326 les coordonnées et fabrique les grilles nécessaires en donnant en sortie les ou et quoi utilisés pour ttm}
\usage{
iso_ouetquoi_4326(
  ou,
  quoi,
  res_ou,
  res_quoi,
  opp_var,
  fun_quoi = "mean",
  resolution = res_quoi,
  rf = 5
)
}
\arguments{
\item{ou}{positions sur lesquelles sont calculés les accessibilités. Si NULL, un raster est défini par défaut.}

\item{quoi}{sf de variables numériques de différentes aménités, qui vont être agrégées au sein d'isochrones.}

\item{res_ou}{resolution des positions de départ.}

\item{res_quoi}{resolution des amenités/opportunités.}

\item{opp_var}{liste des variables d'aménités et d'opportunités.}

\item{fun_quoi}{fonction d'aggrégation des aménités/opportunités à la résolution demandée.}

\item{resolution}{resolution finale. Par défaut, identique à la résolution des aménités/opportunités.}

\item{rf}{je ne sais pas trop.}
}
\description{
construit la grille pour projetter le résultat en crs 4326
\code{iso_ouetquoi} projette sur 4326 les coordonnées et fabrique les grilles nécessaires en donnant en sortie les ou et quoi utilisés pour ttm
}
