% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{iso_split_ou}
\alias{iso_split_ou}
\title{produit les macro-secteurs sur la grille pour compartimenter les calculs.
/code{iso_split_ou} produit les groupes de positions en vue de compartimenter les calculs.}
\usage{
iso_split_ou(ou, quoi, chunk = NULL, routing, tmax = 60)
}
\arguments{
\item{ou}{positions sur lesquelles sont calculés les accessibilités. Si NULL, un raster est défini par défaut.}

\item{quoi}{sf de variables numériques de différentes aménités, qui vont être agrégées au sein d'isochrones.}

\item{chunk}{pas sûr}

\item{routing}{information sur le système de routing.}

\item{tmax}{temps maximal pour les isochrones.}
}
\description{
produit les macro-secteurs sur la grille pour compartimenter les calculs.
/code{iso_split_ou} produit les groupes de positions en vue de compartimenter les calculs.
}
