% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{iso_ttm}
\alias{iso_ttm}
\title{calcul du temps de trajet selon le système de routing choisi.
/code{iso_ttm} calcule un temps de trajets}
\usage{
iso_ttm(o, d, tmax, routing)
}
\arguments{
\item{o}{origine}

\item{d}{destination}

\item{tmax}{temps max pour le trajet}

\item{routing}{système de routing}
}
\description{
calcul du temps de trajet selon le système de routing choisi.
/code{iso_ttm} calcule un temps de trajets
}
