% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{minimax_euclid}
\alias{minimax_euclid}
\title{définition de la distance euclidienne.
/code{minimax_euclid} renvoie la distance euclidienne}
\usage{
minimax_euclid(from, to, dist)
}
\arguments{
\item{from}{point de départ}

\item{to}{point d'arrivée}

\item{dist}{plafond de distance}
}
\description{
définition de la distance euclidienne.
/code{minimax_euclid} renvoie la distance euclidienne
}
