% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobpro.R
\name{mobpro2idINS}
\alias{mobpro2idINS}
\title{paires de trajets suivant mobpro}
\usage{
mobpro2idINS(c200ze, paires_mobpro)
}
\arguments{
\item{c200ze}{un carroyage (ne seront générés que les carreaux dans c200ze)

c200ze doit avoir une colonne com qui contient le code commune,
une colonne ind pour les individus et emp pour les emplois ciblés
ces colonnes sont produites par le carroyage de l'insee et
les données mobpro couplées aux fichiers fonciers

monpro doit avoir une colonne COMMUNE et une colonne DCLT}

\item{mobpro}{les paires de communes}
}
\value{
un tibble avec une colonne fromidINS et une colonne toidINS
}
\description{
A partir d'un carroyage des individus et de l'emploi, et des paires mobpro
on génére les paires à la résolution c200m
}
