% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{OTP_server}
\alias{OTP_server}
\title{lancement d'un serveur java otp}
\usage{
OTP_server(router = "IDF1", port = 8008, memory = "8G", rep)
}
\arguments{
\item{router}{nom du routeur}

\item{port}{par défaut, 8000}

\item{memory}{taille mémoire du serveur, par défaut 8G}

\item{rep}{répertoire}
}
\description{
lancement d'un serveur java otp
}
