% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_utils.r
\name{plage}
\alias{plage}
\title{Plage de dates d'un GTFS}
\usage{
plage(directory)
}
\arguments{
\item{directory}{Répertoire du GTFS}
}
\value{

}
\description{
Lit un GTFS et renvoie la plage de dates commune
aux différents réseaux du GTFS
}
