% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r5 router.R
\name{r5_di}
\alias{r5_di}
\title{calcul des itinéraires détaillés avec r5r}
\usage{
r5_di(o, d, tmax, routing, dist_only = FALSE)
}
\arguments{
\item{o}{origine}

\item{d}{destination}

\item{tmax}{temps max pour le trajet}

\item{routing}{système de routing}

\item{dist_only}{Ne calcule que les distances}
}
\description{
calcul des itinéraires détaillés avec r5r
}
