% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r5 router.R
\name{r5_ttm}
\alias{r5_ttm}
\title{calcul du temps de trajet avec r5}
\usage{
r5_ttm(o, d, tmax, routing)
}
\arguments{
\item{o}{origine}

\item{d}{destination}

\item{tmax}{temps max pour le trajet}

\item{routing}{système de routing}
}
\description{
calcul du temps de trajet avec r5
}
