% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.r
\name{raster_ref}
\alias{raster_ref}
\title{Crée un raster de référence, vide.}
\usage{
raster_ref(data, resolution = 200, crs = 3035)
}
\arguments{
\item{data}{sf.}

\item{resolution}{par défaut, 200.}

\item{crs}{par défaut, 3035 (European Terrestrial Reference System).}
}
\description{
Crée un raster de référence, vide.
}
