% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{routing_setup_euc}
\alias{routing_setup_euc}
\title{setup du système de routing euclidien}
\usage{
routing_setup_euc(mode = "WALK", speed = 5)
}
\arguments{
\item{mode}{mode de transport, par défaut, "WALK".}

\item{speed}{vitesse de déplacement, par défaut, 5km/h.}
}
\description{
setup du système de routing euclidien
}
