% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{routing_setup_osrm}
\alias{routing_setup_osrm}
\title{setup du système de routing osrm}
\usage{
routing_setup_osrm(server = 5000, profile = "driving", future = TRUE)
}
\arguments{
\item{server}{port du serveur osrm, par défaut 5000.}

\item{profile}{mode de transport, par défaut "driving".}

\item{future}{calcul parallele, par défaut TRUE.}
}
\description{
setup du système de routing osrm
}
