% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{routing_setup_otpv1}
\alias{routing_setup_otpv1}
\title{setup du système de routing otp}
\usage{
routing_setup_otpv1(
  router,
  port = 8000,
  memory = "8G",
  rep,
  date = "12-17-2019 8:00:00",
  mode = c("WALK", "TRANSIT"),
  max_walk_time = 30,
  precisionMeters = 50
)
}
\arguments{
\item{router}{info du serveur otp.}

\item{port}{par défaut, 8000.}

\item{memory}{par défaut, 8G.}

\item{rep}{chemin du repertoire.}

\item{date}{date des trajets.}

\item{mode}{mode de transit, par défaut c("WALK", "TRANSIT").}

\item{max_walk_time}{temps maximal à pied}

\item{precisionMeters}{précision demandée au serveur, par défaut 50m.}
}
\description{
setup du système de routing otp
}
