% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{routing_setup_r5}
\alias{routing_setup_r5}
\title{setup du système de routing de r5}
\usage{
routing_setup_r5(
  path,
  date = "17-12-2019 8:00:00",
  mode = c("WALK", "TRANSIT"),
  montecarlo = 1L,
  max_walk_dist = Inf,
  time_window = 1L,
  percentiles = 50L,
  walk_speed = 5,
  bike_speed = 12,
  max_lts = 2,
  max_rides = 5L,
  use_elevation = FALSE,
  overwrite = FALSE,
  n_threads = 4L,
  jMem = "12G",
  di = FALSE,
  quick_setup = FALSE
)
}
\arguments{
\item{path}{path}

\item{date}{date}

\item{mode}{mode de transport, par défaut c("WALK", "TRANSIT")}

\item{montecarlo}{par défaut, 1.}

\item{max_walk_dist}{distance maximale à pied}

\item{time_window}{par défaut, 1. fenêtre pour l'heure de départ.}

\item{percentiles}{par défaut, 50.}

\item{walk_speed}{vitesse piéton}

\item{bike_speed}{vitesse vélo}

\item{max_rides}{nombre de changements de transport.}

\item{n_threads}{nombre de calcul simultané}

\item{jMem}{taille mémoire vive}

\item{di}{renvoie des itinéraires détaillés (distance, nombre de branche) en perdant le montecarlo}

\item{quick_setup}{par défaut, FALSE}
}
\description{
setup du système de routing de r5
}
