% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{safe_r5_ttm}
\alias{safe_r5_ttm}
\title{wrapper pour travel_time_matrix}
\usage{
safe_r5_ttm(...)
}
\arguments{
\item{...}{cf r5r::travel_time_matrix}
}
\description{
wrapper pour travel_time_matrix
}
