% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{select_ancres}
\alias{select_ancres}
\title{sélectionne les macro-secteurs pour le calcul de trajets.}
\usage{
select_ancres(s_ou, k, routing)
}
\arguments{
\item{s_ou}{??}

\item{k}{??}

\item{routing}{??}
}
\description{
sélectionne les macro-secteurs pour le calcul de trajets.
}
