% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil_index.R
\name{theil_index}
\alias{theil_index}
\title{Theil Index}
\usage{
theil_index(
  accessibility_data,
  sociodemographic_data,
  opportunity,
  population,
  socioeconomic_groups,
  group_by = character(0)
)
}
\arguments{
\item{accessibility_data}{A data frame. The accessibility levels whose
inequality should be calculated. Must contain the columns \code{id} and any
others specified in \code{opportunity}.}

\item{sociodemographic_data}{A data frame. The distribution of
sociodemographic characteristics of the population in the study area cells.
Must contain the columns \code{id} and any others specified in \code{population}.}

\item{opportunity}{A string. The name of the column in \code{accessibility_data}
with the accessibility levels to be considerend when calculating inequality
levels.}

\item{population}{A string. The name of the column in \code{sociodemographic_data}
with the number of people in each cell. Used to weigh accessibility levels
when calculating inequality.}

\item{socioeconomic_groups}{A string. The name of the column in
\code{sociodemographic_data} whose values identify the socioeconomic groups that
should be used to calculate the between- and within-groups inequality
levels.}

\item{group_by}{A \code{character} vector. When not \code{character(0)} (the default),
indicates the \code{accessibility_data} columns that should be used to group the
inequality estimates by. For example, if \code{accessibility_data} includes a
\code{scenario} column that identifies distinct scenarios that each
accessibility estimates refer to (e.g. before and after a transport policy
intervention), passing \code{"scenario"} to this parameter results in inequality
estimates grouped by scenario.}
}
\value{
A data frame containing the inequality estimates for the study area.
}
\description{
Theil Index mannn
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data_dir <- system.file("extdata", package = "accessibility")
travel_matrix <- readRDS(file.path(data_dir, "travel_matrix.rds"))
land_use_data <- readRDS(file.path(data_dir, "land_use_data.rds"))

access <- cumulative_cutoff(
  travel_matrix,
  land_use_data,
  cutoff = 30,
  opportunity = "jobs",
  travel_cost = "travel_time"
)

ti <- theil_index(
  access,
  sociodemographic_data = land_use_data,
  opportunity = "jobs",
  population = "population",
  socioeconomic_groups = "income_decile"
)
ti
\dontshow{\}) # examplesIf}
}
\references{
\insertAllCited{}
}
\seealso{
Other inequality: 
\code{\link{concentration_index}()},
\code{\link{gini_index}()},
\code{\link{palma_ratio}()},
\code{\link{theil_index2}()}
}
\concept{inequality}
