% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_closest.R
\name{time_to_closest}
\alias{time_to_closest}
\title{Minimum travel cost to closest N number of opportunities}
\usage{
time_to_closest(
  data,
  opportunity_col,
  n_opportunities = 1,
  travel_cost_col = "travel_time",
  by_col
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel cost matrix in long format,
containing the columns with origin ids, destination ids, the travel cost
between each origin-destination pair (this can be either in  terms of
travel time, distance or monetary costs), and number of population and
opportunities located in each origin and destination.}

\item{opportunity_col}{A string. The name of the column in \code{land_use_data}
with the number of opportunities/resources/services to be considered when
calculating accessibility levels.}

\item{n_opportunities}{A \code{numeric} value with the minimum N number of
opportunities that should be considered. Defaults to \code{1}}

\item{travel_cost_col}{A string. The name of the column in \code{travel_matrix}
with the travel cost between origins and destinations. Defaults to
\code{"travel_time"}.}

\item{by_col}{Either \code{NULL} (the default) or a \code{character} vector. When not
\code{NULL}, indicates the \code{travel_matrix} columns that should be used to group
the accessibility estimates by. For example, if \code{travel_matrix} includes a
departure time column, that specifies the departure time of each entry in
the data frame, passing \code{"departure_time"} to this argument results in
accessibility estimates grouped by origin and by departure time.}
}
\value{
A \code{data.table} object indicating for each origin the travel time to
the closest opportunity and the id of the destination where it is
located.
}
\description{
The function calculates the minimum travel cost to closest N number of
opportunities.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df <- time_to_closest(data = ttm,
                     opportunity_col = 'schools',
                     n_opportunities = 1,
                     travel_cost_col = 'travel_time',
                     by_col = 'from_id')
head(df)

df <- time_to_closest(data = ttm,
                     opportunity_col = 'schools',
                     n_opportunities = 2,
                     travel_cost_col = 'travel_time',
                     by_col = 'from_id')
head(df)

}
\concept{Minimum travel cost}
