% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{ttm_fromfile}
\alias{ttm_fromfile}
\title{Repackage la matrice de distance du disque}
\usage{
ttm_fromfile(dir)
}
\arguments{
\item{dir}{le répertoire où se trouve les fichiers}
}
\value{
un data.table avec les paires o-d et les temps en fonction du mode
}
\description{
A partir de l'output de \code{\link{iso_accessibilite}} en mode \code{ttm_out=TRUE}
calcule les idINS et fabrique un data.table associant les points de départ (from) à ceux de destination (to)
}
