% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{ttm_idINS}
\alias{ttm_idINS}
\title{Repackage la matrice de distance}
\usage{
ttm_idINS(ttm, resolution = 200)
}
\arguments{
\item{ttm}{l'output de \code{\link{iso_accessibilite}}}

\item{resolution}{la résolution par défaut 200m}
}
\value{
un data.table avec les paires o-d et les temps en fonction du mode
}
\description{
A partir de l'output de \code{\link{iso_accessibilite}} en mode \code{ttm_out=TRUE}
calcule les idINS et fabrique un data.table associant les points de départ (from) à ceux de destination (to)
}
