% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{ttm_on_closest}
\alias{ttm_on_closest}
\title{renvoie les matrices de distances, vides si éloignés.}
\usage{
ttm_on_closest(ppou, s_ou, quoi, ttm_0, les_ou, tmax, routing, grdeou)
}
\arguments{
\item{ppou}{??}

\item{s_ou}{??}

\item{quoi}{opportunités}

\item{ttm_0}{??}

\item{les_ou}{points de départ}

\item{tmax}{??}

\item{routing}{info de routing}

\item{grdeou}{??}
}
\description{
renvoie les matrices de distances, vides si éloignés.
}
