% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_utils.r
\name{vmaxmode}
\alias{vmaxmode}
\title{définition des vitesses moyennes selon le mode de transport.}
\usage{
vmaxmode(mode)
}
\arguments{
\item{mode}{mode de transport ("TRANSIT", "RAIL", "BUS", "CAR", "BIKE", "WALK")}
}
\description{
définition des vitesses moyennes selon le mode de transport.
}
