% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\docType{package}
\name{accessibility}
\alias{accessibility}
\alias{_PACKAGE}
\alias{accessibility-package}
\title{accessibility: Transport accessibility metrics}
\description{
Transport accessibility metrics
}
\section{Usage}{

Please check the vignettes for more on the package usage:
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeaGIT/accessibility}
  \item Report bugs at \url{https://github.com/ipeaGIT/accessibility/issues}
}

}
\author{
\strong{Maintainer}: Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})

Authors:
\itemize{
  \item Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})
}

Other contributors:
\itemize{
  \item Ipea - Institute for Applied Economic Research [copyright holder, funder]
}

}
\keyword{internal}
