% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_cutoff.R
\name{cumulative_time_cutoff}
\alias{cumulative_time_cutoff}
\title{Cumulative access based on maximum travel time cutoff}
\usage{
cumulative_time_cutoff(data, opportunity_colname, cutoff, by_colname)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{opportunity_colname}{A \code{string} indicating the name of the column with
data on the opportunities to be considered.}

\item{cutoff}{A \code{numeric} value indicating the maximum travel time considered.}

\item{by_colname}{A \code{string} with the name of the column of origin or
destination that should be considered, indicating whether accessibility
levels should by calculated by each origin (active accessibility) or
destination (passive accessibility).}
}
\value{
A \code{data.table} object.
}
\description{
The function calculates the number of opportunities accessible under a given
travel time cutoff specified by the user.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# Active accessibility: number of schools accessible from each origin
df <- cumulative_time_cutoff(data = ttm,
                            opportunity_colname = 'schools',
                            cutoff = 30,
                            by_colname = 'from_id')
head(df)

# Passive accessibility: number of people that can reach each destination
df <- cumulative_time_cutoff(data = ttm,
                            opportunity_colname = 'population',
                            cutoff = 30,
                            by_colname = 'to_id')
head(df)
}
\seealso{
Other Cumulative access: 
\code{\link{cumulative_time_interval}()},
\code{\link{gravity_access}()}
}
\concept{Cumulative access}
