% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_interval.R
\name{cumulative_time_interval}
\alias{cumulative_time_interval}
\title{Cumulative access based on maximum travel time interval}
\usage{
cumulative_time_interval(
  data,
  opportunity_col,
  interval,
  by_col,
  stat = "mean"
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{opportunity_col}{A \code{string} indicating the name of the column with the
number of opportunities / resources / services to be considered.}

\item{interval}{An \verb{numeric vector} of length 2, indicating the start and end
points of the interval of travel time thresholds to be used.}

\item{by_col}{A \code{string} with the name of the column of origin or
destination that should be considered, indicating whether accessibility
levels should by calculated by each origin (active accessibility) or
destination (passive accessibility).}

\item{stat}{A \code{string} indicating the summary statistic used to aggregate the
accessibility estimates within the time interval. It accepts either \code{median}
(Default) or \code{mean}.}
}
\value{
A \code{data.table} object.
}
\description{
The function calculates the average or median number of opportunities that
can be reached considering multiple minute-by-minute maximum travel time
thresholds within a given travel time interval specified by the user.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df <- cumulative_time_interval(data = ttm,
                              opportunity_col = 'schools',
                              interval = c(20, 30),
                              by_col = 'from_id',
                              stat ='mean')
head(df)

}
\seealso{
Other Cumulative access: 
\code{\link{cumulative_time_cutoff}()}
}
\concept{Cumulative access}
