% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_interval.R
\name{cumulative_time_interval}
\alias{cumulative_time_interval}
\title{Cumulative access based on time interval}
\usage{
cumulative_time_interval(
  data,
  opportunity_colname,
  start,
  end,
  by_colname,
  stat = "mean"
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin, destination, travel time
from origin to destination, and number of opportunities in destination
locations.}

\item{opportunity_colname}{A \code{string} indicating the name of the column with
data on the number of opportunities to be considered.}

\item{start}{An \code{integer} indicating the \code{start} point of the travel time
interval.}

\item{end}{An \code{integer} indicating the \code{end} point of the travel time interval.}

\item{by_colname}{A \code{string} with the name of the column of origin or
destination that should be considered, indicating whether accessibility
levels should by calculated by each origin (active accessibility) or
destination (passive accessibility).}

\item{stat}{A \code{string} indicating the summary statistic used to aggregate the
accessibility estimates within the time interval. It accepts either \code{median}
(Default) or \code{mean}.}
}
\value{
A \code{data.table} object.
}
\description{
The function calculates the average or median number of opportunities
accessible within a travel time interval specified by the user.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_poa.csv", package = "accessibility")
ttm <- read.csv(data_path)

df <- cumulative_time_interval(data = ttm,
                              opportunity_colname = 'schools',
                              start = 20,
                              end = 30,
                              by_colname = 'from_id',
                              stat ='median')
head(df)

}
\seealso{
Other Cumulative access: 
\code{\link{cumulative_time_cutoff}()}
}
\concept{Cumulative access}
