% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_binary.R
\name{decay_binary}
\alias{decay_binary}
\title{Binary (a.k.a. step) decay function}
\usage{
decay_binary(cutoff)
}
\arguments{
\item{cutoff}{A \code{numeric}. A number indicating the travel cost cutoff.}
}
\value{
A \code{function} that takes a generic travel cost (\code{numeric}) as an input
and returns an impedance factor (\code{numeric}).
}
\description{
Returns a binary impedance function (frequently used to calculate cumulative
opportunities measures) to be used inside accessibility calculating
functions.

This function is generic over any kind of numeric travel cost,
such as distance, time and money.
}
\examples{
impedance <- decay_binary(cutoff = 30)

impedance(t_ij = 20)

impedance(t_ij = 35)

}
\seealso{
Other Impedance functions: 
\code{\link{decay_exponential}()},
\code{\link{decay_linear}()},
\code{\link{decay_power}()}
}
\concept{Impedance functions}
