% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr router.R
\name{dgr_distances_by_com}
\alias{dgr_distances_by_com}
\title{Full distance}
\usage{
dgr_distances_by_com(
  idINSes,
  com2com,
  routeur,
  path = "dgr",
  overwrite = TRUE,
  parallel = TRUE,
  clusterize = FALSE
)
}
\arguments{
\item{idINSes}{une table d'idINS, avec les colonnes idINS, from (lgl), to (lgl) COMMUNE, DCLT}

\item{com2com}{une table donnant les paires de relation}

\item{routeur}{un routeur dodgr, généré par routing_setup_dodgr}

\item{parallel}{fait le calcul en parallèle (défaut TRUE)}

\item{clusterize}{regroupe les communes pour minimiser le temps de caclul (défaut FALSE)
pour chaque paire COMMUNE, DCLT le routage est fait sur le produit des
from et to indiqués pour cette paire}
}
\value{
un data.table, tous calculs faits
}
\description{
Alternative à iso acessibilité, plus simple
ne renvoie que les distances, temps et dzplus
exécute le calcul par COMMUNE
certains batchs sont assez petits mais relativement efficace au total
}
