% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{euc_ttm}
\alias{euc_ttm}
\title{calcul du temps de trajet avec une distance euclidienne}
\usage{
euc_ttm(o, d, tmax, routing)
}
\arguments{
\item{o}{origine}

\item{d}{destination}

\item{tmax}{temps max pour le trajet}

\item{routing}{système de routing}
}
\description{
calcul du temps de trajet avec une distance euclidienne
}
