% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_2sfca.R
\name{fca_2sfca}
\alias{fca_2sfca}
\title{2SFCA two-step floating catchment area}
\usage{
fca_2sfca(
  data,
  orig_col,
  dest_col,
  population_col,
  opportunity_col,
  decay_function
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{orig_col}{A \code{string} with the name of the column of origin ids.}

\item{dest_col}{A \code{string} with the name of the column of destination ids.}

\item{population_col}{A \code{string} with the name of the column of origin with
population count.}

\item{opportunity_col}{A \code{string} with the name of the column of destination
with  the number of opportunities / resources / services.}

\item{decay_function}{A \code{fuction} that converts travel cost into and impedance
factor used to weigth opportunities. For convinence, the package currently
includes the following functions: \code{\link[=decay_bineary]{decay_bineary()}}, \code{\link[=decay_linear]{decay_linear()}} and
\code{\link[=decay_exponential]{decay_exponential()}}. See the documentation of each function for more
details.}
}
\value{
A \code{numeric} estimate of accessibility.
}
\description{
Calculates accessibility levels using the two-step floating catchment area
(2SFCA) proposed by Luo and Wang (2003). The 2SFCA metric was the first
accessibility metric in the the floating catchment area family.
}
\details{
References:
\itemize{
\item Luo, W., & Wang, F. (2003). Measures of spatial accessibility to health
care in a GIS environment: synthesis and a case study in the Chicago region.
Environment and planning B: planning and design, 30(6), 865-884. \doi{10.1068/b29120}.
}
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# 2SFCA with a step decay function
df <- fca_2sfca(data = ttm,
              orig_col = 'from_id',
              dest_col = 'to_id',
              opportunity_col = 'jobs',
              population_col = 'population',
              decay_function = decay_linear(cutoff = 50)
              )
head(df)

# 2SFCA with an exponential decay function
df2 <- fca_2sfca(data = ttm,
        orig_col = 'from_id',
        dest_col = 'to_id',
        opportunity_col = 'jobs',
        population_col = 'population',
        decay_function = decay_exponential(decay_value = 0.5)
        )

head(df2)

}
\seealso{
Other Floating catchment area: 
\code{\link{fca_bfca}()},
\code{\link{floating_catchment_area}()}
}
\concept{Floating catchment area}
