% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_bfca.R
\name{fca_bfca}
\alias{fca_bfca}
\title{BFCA balanced floating catchment area}
\usage{
fca_bfca(
  data,
  orig_col,
  dest_col,
  population_col,
  opportunity_col,
  decay_function,
  cutoff = NULL,
  decay_value = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{orig_col}{A \code{string} with the name of the column of origin ids.}

\item{dest_col}{A \code{string} with the name of the column of destination ids.}

\item{population_col}{A \code{string} with the name of the column of origin with
population count.}

\item{opportunity_col}{A \code{string} with the name of the column of destination
with  the number of opportunities / resources / services.}

\item{decay_function}{A string. Which decay function to use when calculating
accessibility. One of step, exponential, fixed_exponential, linear
or logistic. Please see the details to understand how each
alternative works and how they relate to the \code{cutoffs} and
\code{decay_value} parameters.}

\item{cutoff}{A numeric vector. This parameter has different effects for each
decay function: it indicates the cutoff times in minutes when
calculating cumulative opportunities accessibility with the
\code{step} function...}

\item{decay_value}{A number. Extra parameter to be passed to the selected
\code{decay_function}. Has no effects when \code{decay_function} is either
\code{step} or \code{exponential}.}
}
\value{
A \code{numeric} estimate of accessibility.
}
\description{
Calculates accessibility levels using the balanced floating catchment area
(BFCA) measure proposed by Paez et al. (2019). The BFCS metric calculate
accessibility accouting for the competition of resources and simultaneously
correcting for issues of inflation of demand and service levels that are
present in other accessibility metrics in the floating catchment area family.
}
\details{
The balanced floating catchment area (BFCA) measure was originally proposed
by Paez et al. (2019) and
\itemize{
\item Paez, A., Higgins, C. D., & Vivona, S. F. (2019). Demand and level of
service inflation in Floating Catchment Area (FCA) methods. Plos one, 14(6),
e0218773. \doi{10.1371/journal.pone.0218773}
\item Pereira, R. H., Braga, C. K. V., Servo, L. M., Serra, B., Amaral, P.,
Gouveia, N., & Paez, A. (2021). Geographic access to COVID-19 healthcare in
Brazil using a balanced float catchment area approach. Social Science &
Medicine, 273. \doi{10.1016/j.socscimed.2021.113773}
}
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# BFCA with a step decay function
df <- fca_bfca(data = ttm,
              orig_col = 'from_id',
              dest_col = 'to_id',
              opportunity_col = 'jobs',
              population_col = 'population',
              decay_function = 'step',
              cutoff = 30)
head(df)

df2 <- fca_bfca(data = ttm,
               orig_col = 'from_id',
               dest_col = 'to_id',
               opportunity_col = 'jobs',
               population_col = 'population',
               decay_function = 'negative_exponential',
               decay_value = 0.5)

head(df2)

}
\concept{Floating catchment area}
