% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_bfca.R
\name{fca_bfca}
\alias{fca_bfca}
\title{BFCA balanced floating catchment area}
\usage{
fca_bfca(
  data,
  orig_col,
  dest_col,
  population_col,
  opportunity_col,
  decay_function
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{orig_col}{A \code{string} with the name of the column of origin ids.}

\item{dest_col}{A \code{string} with the name of the column of destination ids.}

\item{population_col}{A \code{string} with the name of the column of origin with
population count.}

\item{opportunity_col}{A \code{string} with the name of the column of destination
with  the number of opportunities / resources / services.}

\item{decay_function}{A \code{fuction} that converts travel cost (t_ij) into an
impedance factor used to weight opportunities. For convinence, the package
currently includes the following functions: \code{decay_bineary},
\code{decay_exponential}, \code{decay_exponential} and \code{decay_power.}.
See the documentation of each function for more details.}
}
\value{
A \code{numeric} estimate of accessibility.
}
\description{
Calculates accessibility levels using the balanced floating catchment area
(BFCA) measure proposed by Paez et al. (2019). The BFCS metric calculate
accessibility accouting for the competition of resources and simultaneously
correcting for issues of inflation of demand and service levels that are
present in other accessibility metrics in the floating catchment area family.
}
\details{
The balanced floating catchment area (BFCA) measure was originally proposed
by Paez et al. (2019) and
\itemize{
\item Paez, A., Higgins, C. D., & Vivona, S. F. (2019). Demand and level of
service inflation in Floating Catchment Area (FCA) methods. Plos one, 14(6),
e0218773. \doi{10.1371/journal.pone.0218773}
\item Pereira, R. H., Braga, C. K. V., Servo, L. M., Serra, B., Amaral, P.,
Gouveia, N., & Paez, A. (2021). Geographic access to COVID-19 healthcare in
Brazil using a balanced float catchment area approach. Social Science &
Medicine, 273. \doi{10.1016/j.socscimed.2021.113773}
}
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# BFCA with a step decay function
df <- fca_bfca(data = ttm,
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = decay_binary(cutoff = 50)
       )

head(df)

df2 <- fca_bfca(data = ttm,
        orig_col = 'from_id',
        dest_col = 'to_id',
        opportunity_col = 'jobs',
        population_col = 'population',
        decay_function = decay_exponential(decay_value = 0.5)
        )

head(df2)

}
\seealso{
Other Floating catchment area: 
\code{\link{fca_2sfca}()},
\code{\link{floating_catchment_area}()}
}
\concept{Floating catchment area}
