% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_model.R
\name{gravity_access}
\alias{gravity_access}
\title{Gravity-based accessibility measures}
\usage{
gravity_access(
  data,
  opportunity_col,
  decay_function,
  travel_cost_col = "travel_time",
  by_col
)
}
\arguments{
\item{data}{A \code{data.frame} with a travel cost matrix in long format,
containing the columns with origin ids, destination ids, the travel cost
between each origin-destination pair (this can be either in  terms of
travel time, distance or monetary costs), and number of population and
opportunities located in each origin and destination.}

\item{opportunity_col}{A string. The name of the column in \code{land_use_data}
with the number of opportunities/resources/services to be considered when
calculating accessibility levels.}

\item{decay_function}{A \code{fuction} that converts travel cost (t_ij) into an
impedance factor used to weight opportunities. For convenience, the package
currently includes the following functions: \code{decay_bineary},
\code{decay_exponential}, \code{decay_exponential} and \code{decay_power}.
See the documentation of each decay function for more details.}

\item{travel_cost_col}{A string. The name of the column in \code{travel_matrix}
with the travel cost between origins and destinations. Defaults to
\code{"travel_time"}.}

\item{by_col}{Either \code{NULL} (the default) or a \code{character} vector. When not
\code{NULL}, indicates the \code{travel_matrix} columns that should be used to group
the accessibility estimates by. For example, if \code{travel_matrix} includes a
departure time column, that specifies the departure time of each entry in
the data frame, passing \code{"departure_time"} to this argument results in
accessibility estimates grouped by origin and by departure time.}
}
\value{
A \code{data.table} object.
}
\description{
The function calculates gravity-based accessibility measures. Accessibility
can be calculated using multiple alternative decay functions passed through
the \code{decay_function} parameter.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df_linear <- gravity_access(data = ttm,
                           opportunity_col = 'schools',
                           decay_function = decay_linear(cutoff = 50),
                           travel_cost_col = 'travel_time',
                           by_col = 'from_id'
                           )

head(df_linear)

df_exp <- gravity_access(data = ttm,
                        opportunity_col = 'schools',
                        decay_function = decay_exponential(decay_value = 0.5),
                        travel_cost_col = 'travel_time',
                        by_col = 'from_id'
                        )
head(df_exp)

}
\concept{Gravity-based accessibility}
