% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_model.R
\name{gravity_access}
\alias{gravity_access}
\title{Gravity-based accessibility measures}
\usage{
gravity_access(data, opportunity_col, by_col, decay_function)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{opportunity_col}{A \code{string} indicating the name of the column with
data on the opportunities to be considered.}

\item{by_col}{A \code{string} with the name of the column of origin or
destination that should be considered, indicating whether accessibility
levels should by calculated by each origin (active accessibility) or
destination (passive accessibility).}

\item{decay_function}{A \code{fuction} that converts travel cost (t_ij) into an
impedance factor used to weight opportunities. For convinence, the package
currently includes the following functions: \code{decay_bineary},
\code{decay_exponential}, \code{decay_exponential} and \code{decay_power.}.
See the documentation of each function for more details.}
}
\value{
A \code{data.table} object.
}
\description{
The function calculates gravity-based accessibility measures...
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df_linear <- gravity_access(data = ttm,
                           opportunity_col = 'schools',
                           by_col = 'from_id',
                           decay_function = decay_linear(cutoff = 50)
                           )

head(df_linear)

df_exp <- gravity_access(data = ttm,
                        opportunity_col = 'schools',
                        by_col = 'from_id',
                        decay_function = decay_exponential(decay_value = 0.5)
                        )
head(df_exp)

}
\concept{Gravity-based accessibility}
