% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_system.r
\name{routing_setup_r5}
\alias{routing_setup_r5}
\title{setup du système de routing de r5}
\usage{
routing_setup_r5(
  path,
  date = "17-12-2019 8:00:00",
  mode = c("WALK", "TRANSIT"),
  montecarlo = 10L,
  max_walk_time = Inf,
  time_window = 1L,
  percentiles = 50L,
  walk_speed = 5,
  bike_speed = 12,
  max_lts = 2,
  max_rides = ifelse("TRANSIT" \%in\% mode, 3L, 1L),
  use_elevation = FALSE,
  elevation_tif = NULL,
  elevation = "NONE",
  dfMaxLength = 10,
  breakdown = FALSE,
  overwrite = FALSE,
  n_threads = 4L,
  max_rows = 5000,
  jMem = "12G",
  di = FALSE,
  quick_setup = FALSE
)
}
\arguments{
\item{path}{path Chemin d'accès au dossier}

\item{date}{date Date où seront simulées les routes}

\item{mode}{mode de transport, par défaut c("WALK", "TRANSIT"), voir r5r ou r5 pour les autres modes}

\item{montecarlo}{nombre de tirages montecarlo par minutes de time_windows par défaut 10}

\item{max_walk_time}{temps maximal à pied}

\item{time_window}{par défaut, 1. fenêtre pour l'heure de départ en minutes}

\item{percentiles}{par défaut, 50., retourne les percentiles de temps de trajet (montecarlo)}

\item{walk_speed}{vitesse piéton}

\item{bike_speed}{vitesse vélo}

\item{max_lts}{stress maximal à vélo (de 1 enfant à 4 toutes routes), 2 par défaut}

\item{max_rides}{nombre maximal de changements de transport}

\item{use_elevation}{le routing est effectué en utilisant l'information de dénivelé. Pas sûr que cela fonctionne.}

\item{elevation_tif}{nom du fichier raster (WGS84) des élévations en mètre, en passant ce paramètre, on calcule le dénivelé positif.
elevatr::get_elev_raster est un bon moyen de le générer. Fonctionne même si on n'utilise pas les élévations dans le routing}

\item{elevation}{méthode de clcul (NONE, TOBLER, )}

\item{dfMaxLength}{longueur en mètre des segments pour la discrétization}

\item{overwrite}{Regénére le network.dat même si il est présent}

\item{n_threads}{nombre de threads}

\item{max_rows}{nombre maximale de lignes passées à detailled_itirenaries}

\item{jMem}{taille mémoire vive, plus le nombre de threads est élevé, plus la mémoire doit être importante}

\item{di}{renvoie des itinéraires détaillés (distance, nombre de branche) en perdant le montecarlo et au prix d'une plus grande lenteur}

\item{quick_setup}{dans le cas où le core existe déjà, il n'est pas recréer, plus rapide donc, par défaut, FALSE}
}
\description{
Cette fonction est utlisée pour fabriquer le "core" r5 qui est ensuite appelé pour faire le routage.
Elle utilise le r5.....jar téléchargé ou le télécharge si nécessaire.
Un dossier contenant les informations nécessaires doit être passé en parmètre.
Ce dossier contient au moins un pbf (OSM), des fichiers GTFS (zip), des fichiers d'élévation (.tif)
Si iol contient un network.dat, celui ci sera utilisé sans reconstruire le réseau
}
