% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_closest.R
\name{time_to_closest}
\alias{time_to_closest}
\title{Minimum travel time to closest opportunities}
\usage{
time_to_closest(data, opportunity_col, by_col, n_opportunities = 1)
}
\arguments{
\item{data}{A \code{data.frame} with a travel time matrix in long format,
containing the at least the columns of origin \code{from_id}, destination \code{to_id},
travel time \code{travel_time} from origin to destination, and number of
opportunities in destination locations.}

\item{opportunity_col}{A \code{string} indicating the name of the column with
data on the opportunities to be considered.}

\item{by_col}{A \code{string} with the name of the column of origin or
destination that should be considered, indicating whether accessibility
levels should by calculated by each origin (active accessibility) or
destination (passive accessibility).}

\item{n_opportunities}{A \code{numeric} value with the mininum N number of
opportunities that should be considered. Defaults to \code{1}}
}
\value{
A \code{data.table} object indicating for each origin the travel time to
the closest opportunity and the id of the destination where it is located.
}
\description{
The function calculates the minimum travel time to closest N number of
opportunities.
}
\examples{
library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df <- time_to_closest(data = ttm,
                     opportunity_col = 'schools',
                     by_col = 'from_id',
                     n_opportunities = 1)
head(df)

df <- time_to_closest(data = ttm,
                     opportunity_col = 'schools',
                     by_col = 'from_id',
                     n_opportunities = 2)
head(df)

}
\concept{Minimum travel time}
